feat: likes implementation started (#13)
This pull request introduces a new feature for handling user likes,
along with some refactoring and improvements to existing code. The most
important changes include the addition of a `LikesController`,
`LikesModel`, and `LikesRouter`, modifications to the `UsersController`,
and updates to the database schema.

### New Feature: User Likes

*
[`Backend/src/Controllers/LikesController.js`](diffhunk://#diff-ee8e51ce3c5b126c58c7d8f539bf2f2706689774320dc12f962766b2f81f2a63R1-R74):
Added a new `LikesController` class to handle the logic for liking and
unliking users.
*
[`Backend/src/Models/LikesModel.js`](diffhunk://#diff-a9a64b864cb815fc82b12393c9fda3a3ff2503e00d72d6886f75d27f66c9bb1fR1-R11):
Created a new `LikesModel` class to interact with the likes table in the
database.
*
[`Backend/src/Routes/LikesRouter.js`](diffhunk://#diff-346617f4818128bfc7f6fd90897c9d22188c618bece583ad73e27c4ce57dde4fR1-R16):
Added a new `LikesRouter` class to define the routes for handling likes.
*
[`Backend/src/Core/App.js`](diffhunk://#diff-cf0127a62c9d06bba563b1261912b42af5cd41f612138f2a7625a58de01479acR18):
Integrated the `LikesRouter` into the main application router.
[[1]](diffhunk://#diff-cf0127a62c9d06bba563b1261912b42af5cd41f612138f2a7625a58de01479acR18)
[[2]](diffhunk://#diff-cf0127a62c9d06bba563b1261912b42af5cd41f612138f2a7625a58de01479acR65)

### Refactoring and Improvements

*
[`Backend/src/Controllers/UsersController.js`](diffhunk://#diff-56f018579659adfeaf94449113cc636c51fe4316620354c4bf5d89439d20323bL9-R11):
Refactored methods to use `returnErrorStatus` for error handling and
renamed `saveAsView` to `saveView`.
[[1]](diffhunk://#diff-56f018579659adfeaf94449113cc636c51fe4316620354c4bf5d89439d20323bL9-R11)
[[2]](diffhunk://#diff-56f018579659adfeaf94449113cc636c51fe4316620354c4bf5d89439d20323bL371-R371)
[[3]](diffhunk://#diff-56f018579659adfeaf94449113cc636c51fe4316620354c4bf5d89439d20323bL412-R420)
[[4]](diffhunk://#diff-56f018579659adfeaf94449113cc636c51fe4316620354c4bf5d89439d20323bL432-R438)

### Database Schema Update

*
[`Docker/PostgreSQL/init/init.sql`](diffhunk://#diff-1cad65cb0ae9b2f586380e45ebf3748c081e43d844123bd5d091a971867a7a5fL49-R51):
Modified the `likes` table schema to change the column name from
`liked_to` to `liked`.

### Utility and Status Messages

*
[`Backend/src/Utils/StatusMessage.js`](diffhunk://#diff-5336c082fa6c307689cbe5da49cd097a461330c534c4167a1b5edef9f9095be6R67-R69):
Added new status messages for the likes feature.

These changes collectively introduce the functionality for users to like
and unlike other users, improve error handling, and ensure the codebase
remains clean and maintainable.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>