refactor: name change for views_history related names (#16)
This pull request includes several important changes to the
`UsersController` and related models to rename and refactor the visit
history functionality to views history. The changes involve updates to
method names, variable names, and model names to reflect this
refactoring.

Key changes include:

### Refactoring visit history to views history:

*
[`Backend/src/Controllers/UsersController.js`](diffhunk://#diff-56f018579659adfeaf94449113cc636c51fe4316620354c4bf5d89439d20323bL14-R14):
Renamed `visitHistoryModel` to `viewsHistoryModel` and updated method
parameters and variable names to reflect the new terminology.
[[1]](diffhunk://#diff-56f018579659adfeaf94449113cc636c51fe4316620354c4bf5d89439d20323bL14-R14)
[[2]](diffhunk://#diff-56f018579659adfeaf94449113cc636c51fe4316620354c4bf5d89439d20323bL441-R469)
[[3]](diffhunk://#diff-56f018579659adfeaf94449113cc636c51fe4316620354c4bf5d89439d20323bL488-R488)
*
[`Backend/src/Models/ViewsHistoryModel.js`](diffhunk://#diff-00b69d7538e2c19bf4fdc059739d771e415f87f5e8cc4c47c9eb1956eef14cc0L4-R6):
Renamed the file from `VisitHistoryModel.js` to `ViewsHistoryModel.js`
and updated the class and method names accordingly.
[[1]](diffhunk://#diff-00b69d7538e2c19bf4fdc059739d771e415f87f5e8cc4c47c9eb1956eef14cc0L4-R6)
[[2]](diffhunk://#diff-00b69d7538e2c19bf4fdc059739d771e415f87f5e8cc4c47c9eb1956eef14cc0L15-R21)
[[3]](diffhunk://#diff-00b69d7538e2c19bf4fdc059739d771e415f87f5e8cc4c47c9eb1956eef14cc0L36-R37)
*
[`Docker/PostgreSQL/init/init.sql`](diffhunk://#diff-1cad65cb0ae9b2f586380e45ebf3748c081e43d844123bd5d091a971867a7a5fL54-R57):
Updated the table name from `visit_history` to `views_history` and
renamed the columns to match the new terminology.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>