Feat/calculate distance endpoint (#28)
This pull request introduces a new feature to calculate the distance
between two locations. The changes include the addition of a new
controller, router, schema validation, and updates to the main
application to integrate the new functionality.

### New Feature: Distance Calculation

*
[`Backend/src/Controllers/DistanceController.js`](diffhunk://#diff-824b44010d26c38369ac28e682b9041b20b1a669b9c1b795e05f8d927f576beaR1-R17):
Added `DistanceController` class with a `calculateDistance` method to
handle distance calculation requests. This method validates the input
locations and returns the calculated distance.
*
[`Backend/src/Routes/DistanceRouter.js`](diffhunk://#diff-47b1cfc1e374f4c2c64f038a4f2d1d93a81b4f38a0124466ec22550b3c829731R1-R16):
Added `DistanceRouter` class to define the route for the distance
calculation endpoint.
*
[`Backend/src/Schemas/locationsSchema.js`](diffhunk://#diff-97c45b5a02db3344bd07892ae8b4f908878042078ffce955c0bf11520c70c124R1-R62):
Added schema validation for the locations using `zod` library to ensure
the input data is correct.

### Integration with Main Application

*
[`Backend/src/Core/App.js`](diffhunk://#diff-cf0127a62c9d06bba563b1261912b42af5cd41f612138f2a7625a58de01479acR22):
Imported `DistanceRouter` and added the new distance calculation route
to the application.
[[1]](diffhunk://#diff-cf0127a62c9d06bba563b1261912b42af5cd41f612138f2a7625a58de01479acR22)
[[2]](diffhunk://#diff-cf0127a62c9d06bba563b1261912b42af5cd41f612138f2a7625a58de01479acR81)

### Utility Updates

*
[`Backend/src/Utils/StatusMessage.js`](diffhunk://#diff-5336c082fa6c307689cbe5da49cd097a461330c534c4167a1b5edef9f9095be6R98):
Added a new status message for cases where distance calculation cannot
be performed due to missing locations.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>