fix: general bug fixing (#31)
This pull request to the `Backend/src` directory includes changes to
improve error handling and validation in the `BrowserController` and
`eventSchema` files. The most important changes include modifying the
response for no users found and adding a non-empty validation rule for
event titles.

Error handling improvements:

*
[`Backend/src/Controllers/BrowserController.js`](diffhunk://#diff-d613d186e7ef8ed52ed613e4a218da0548b8745bc9c9ca33f88ac1f207787038L23-R23):
Modified the response for no users found to return a 404 status with an
empty JSON object instead of a message.

Validation improvements:

*
[`Backend/src/Schemas/eventSchema.js`](diffhunk://#diff-d800286263e4549dd22622951766d63528f0d5fec8a19bfa2a6f527d16af0b97R22):
Added a `.nonempty` validation rule to ensure the event title cannot be
empty.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>