feat: notification read status (#38)
This pull request introduces several changes to the notifications
system, including adding a read status to notifications and implementing
functionality to mark notifications as read. The most important changes
are summarized below:

### Enhancements to Notifications System:

*
[`Backend/src/Controllers/NotificationsController.js`](diffhunk://#diff-3273a5a76b751e7d1b777b3fd1565697b18919ea73d1909b2cdbbef1f4bb092cR35-R37):
Added `id` and `read` fields to the notification object, and introduced
a new static method `markRead` to update the read status of
notifications.
[[1]](diffhunk://#diff-3273a5a76b751e7d1b777b3fd1565697b18919ea73d1909b2cdbbef1f4bb092cR35-R37)
[[2]](diffhunk://#diff-3273a5a76b751e7d1b777b3fd1565697b18919ea73d1909b2cdbbef1f4bb092cR62-R73)
*
[`Backend/src/Models/NotificationsModel.js`](diffhunk://#diff-9a65fafb5ee89e5b8259fc1cd6caae2c11e984505ade40f2efb399c9a8a20200R8-R26):
Added the `updateReadStatus` method to update the read status of
notifications in the database.
*
[`Backend/src/Routes/NotificationsRouter.js`](diffhunk://#diff-62ba0354429dcf851f2540014def725abeb4a90977f0816836b360f6e7c14436R14-R16):
Added a new PATCH route `/mark-read` to handle marking notifications as
read.
*
[`Backend/src/Sockets/Notifications.js`](diffhunk://#diff-02ca0bbe0b48152fa5c6e7778afa7dee1f9cd07cb1645de05ed922f984484236R41-R43):
Added `id` and `read` fields to the notification payload sent through
sockets.
*
[`Docker/PostgreSQL/init/init.sql`](diffhunk://#diff-1cad65cb0ae9b2f586380e45ebf3748c081e43d844123bd5d091a971867a7a5fR147):
Added a new `read` column to the `notifications` table to store the read
status of notifications.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>