fix: general fixes  (#44)
This pull request includes several changes focused on enhancing the
functionality of the dislike feature, updating the likes controller, and
improving the user interface for profile details. The most important
changes include the addition of the `DislikesModel`, modifications to
the `LikesController` to handle dislikes, and improvements to the
profile details page to show a tooltip with the exact last online time.

Enhancements to dislike functionality:

*
[`Backend/src/Models/DislikesModel.js`](diffhunk://#diff-8eec1d4373f7aad0b6b69760c987f100d2afe670d3eeb9006d701051832a63aaR1-R24):
Added a new model, `DislikesModel`, to manage dislike data, including
methods to check if a user is disliked.
*
[`Docker/PostgreSQL/init/init.sql`](diffhunk://#diff-1cad65cb0ae9b2f586380e45ebf3748c081e43d844123bd5d091a971867a7a5fR84-R91):
Created a new `dislikes` table in the database to store dislike
information.

Updates to `LikesController`:

*
[`Backend/src/Controllers/LikesController.js`](diffhunk://#diff-ee8e51ce3c5b126c58c7d8f539bf2f2706689774320dc12f962766b2f81f2a63R12):
Modified the `handleLike` method to remove a dislike if it exists when a
like is added and added logic to create a dislike entry.
[[1]](diffhunk://#diff-ee8e51ce3c5b126c58c7d8f539bf2f2706689774320dc12f962766b2f81f2a63R12)
[[2]](diffhunk://#diff-ee8e51ce3c5b126c58c7d8f539bf2f2706689774320dc12f962766b2f81f2a63R126-R132)
[[3]](diffhunk://#diff-ee8e51ce3c5b126c58c7d8f539bf2f2706689774320dc12f962766b2f81f2a63R175-R188)

Improvements to notifications:

*
[`Backend/src/Sockets/Notifications.js`](diffhunk://#diff-02ca0bbe0b48152fa5c6e7778afa7dee1f9cd07cb1645de05ed922f984484236R3):
Updated the notifications logic to prevent sending notifications if the
recipient has disliked the notifier.
[[1]](diffhunk://#diff-02ca0bbe0b48152fa5c6e7778afa7dee1f9cd07cb1645de05ed922f984484236R3)
[[2]](diffhunk://#diff-02ca0bbe0b48152fa5c6e7778afa7dee1f9cd07cb1645de05ed922f984484236R26-R32)

User interface enhancements:

*
[`Frontend/src/pages/PublicProfile/ProfileDetails.tsx`](diffhunk://#diff-37a830e558bf77611237ca4b8b74e33521f6711c054ddd7801d7857492e36799R1-R7):
Added a tooltip to display the exact last online time when hovering over
the last seen information.
[[1]](diffhunk://#diff-37a830e558bf77611237ca4b8b74e33521f6711c054ddd7801d7857492e36799R1-R7)
[[2]](diffhunk://#diff-37a830e558bf77611237ca4b8b74e33521f6711c054ddd7801d7857492e36799L20-R42)

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>
Co-authored-by: magnitopic <magnitopic@gmail.com>